Inline test options
