+++ src/Cli/dotnet/Commands/Test/MTP/OutputOptions.cs
@@ -1,0 +1,10 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+namespace Microsoft.DotNet.Cli.Commands.Test;
+
+internal enum OutputOptions
+{
+    Normal,
+    Detailed
+}
