+++ src/Cli/dotnet/Commands/Test/TestCommandDefinition.cs
@@ -6,6 +6,7 @@ using System.CommandLine;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Microsoft.DotNet.Cli.CommandLine;
+using Microsoft.DotNet.Cli.Extensions;
 using Microsoft.TemplateEngine.Cli.Commands;
 using Microsoft.TemplateEngine.Cli.Help;
 using Command = System.CommandLine.Command;
@@ -94,29 +95,126 @@ internal abstract class TestCommandDefinition : Command
 
     public sealed class MicrosoftTestingPlatform : TestCommandDefinition, ICustomHelp
     {
-        public readonly Option<string> ProjectOrSolutionOption = MicrosoftTestingPlatformOptions.ProjectOrSolutionOption;
-        public readonly Option<string> SolutionOption = MicrosoftTestingPlatformOptions.SolutionOption;
-        public readonly Option<string> TestModulesFilterOption = MicrosoftTestingPlatformOptions.TestModulesFilterOption;
-        public readonly Option<string> TestModulesRootDirectoryOption = MicrosoftTestingPlatformOptions.TestModulesRootDirectoryOption;
-        public readonly Option<string> ResultsDirectoryOption = MicrosoftTestingPlatformOptions.ResultsDirectoryOption;
-        public readonly Option<string> ConfigFileOption = MicrosoftTestingPlatformOptions.ConfigFileOption;
-        public readonly Option<string> DiagnosticOutputDirectoryOption = MicrosoftTestingPlatformOptions.DiagnosticOutputDirectoryOption;
-        public readonly Option<int> MaxParallelTestModulesOption = MicrosoftTestingPlatformOptions.MaxParallelTestModulesOption;
-        public readonly Option<int> MinimumExpectedTestsOption = MicrosoftTestingPlatformOptions.MinimumExpectedTestsOption;
+        public readonly Option<string> ProjectOrSolutionOption = new("--project")
+        {
+            Description = CliCommandStrings.CmdProjectOrSolutionDescriptionFormat,
+            HelpName = CliCommandStrings.CmdProjectOrSolutionPathName,
+            Arity = ArgumentArity.ExactlyOne
+        };
+
+        public readonly Option<string> SolutionOption = new("--solution")
+        {
+            Description = CliCommandStrings.CmdSolutionDescription,
+            HelpName = CliCommandStrings.CmdSolutionPathName,
+            Arity = ArgumentArity.ExactlyOne
+        };
+
+        public readonly Option<string> TestModulesFilterOption = new("--test-modules")
+        {
+            Description = CliCommandStrings.CmdTestModulesDescription,
+            HelpName = CliCommandStrings.CmdExpressionName
+        };
+
+        public readonly Option<string> TestModulesRootDirectoryOption = new("--root-directory")
+        {
+            Description = CliCommandStrings.CmdTestModulesRootDirectoryDescription,
+            HelpName = CliCommandStrings.CmdRootPathName,
+        };
+
+        public const string ResultsDirectoryOptionName = "--results-directory";
+
+        public readonly Option<string> ResultsDirectoryOption = new(ResultsDirectoryOptionName)
+        {
+            Description = CliCommandStrings.CmdResultsDirectoryDescription,
+            HelpName = CliCommandStrings.CmdPathToResultsDirectory,
+            Arity = ArgumentArity.ExactlyOne
+        };
+
+        public const string ConfigFileOptionName = "--config-file";
+
+        public readonly Option<string> ConfigFileOption = new(ConfigFileOptionName)
+        {
+            Description = CliCommandStrings.CmdConfigFileDescription,
+            HelpName = CliCommandStrings.CmdConfigFilePath,
+            Arity = ArgumentArity.ExactlyOne
+        };
+
+        public const string DiagnosticOutputDirectoryOptionName = "--diagnostic-output-directory";
+
+        public readonly Option<string> DiagnosticOutputDirectoryOption = new(DiagnosticOutputDirectoryOptionName)
+        {
+            Description = CliCommandStrings.CmdDiagnosticOutputDirectoryDescription,
+            HelpName = CliCommandStrings.CmdDiagnosticOutputDirectoryPath,
+            Arity = ArgumentArity.ExactlyOne
+        };
+
+        public readonly Option<int> MaxParallelTestModulesOption = new("--max-parallel-test-modules")
+        {
+            Description = CliCommandStrings.CmdMaxParallelTestModulesDescription,
+            HelpName = CliCommandStrings.CmdNumberName
+        };
+
+        public readonly Option<int> MinimumExpectedTestsOption = new("--minimum-expected-tests")
+        {
+            Description = CliCommandStrings.CmdMinimumExpectedTestsDescription,
+            HelpName = CliCommandStrings.CmdNumberName
+        };
+
         public readonly Option<string> ArchitectureOption = CommonOptions.ArchitectureOption;
+
         public readonly Option<IReadOnlyDictionary<string, string>> EnvOption = CommonOptions.EnvOption;
+
         public readonly Option<ReadOnlyDictionary<string, string>?> PropertiesOption = CommonOptions.PropertiesOption;
+
         public readonly Option<string> OperatingSystemOption = CommonOptions.OperatingSystemOption;
+
         public readonly Option<string> RuntimeOption = CommonOptions.CreateRuntimeOption(CliCommandStrings.TestRuntimeOptionDescription);
+
         public readonly Option<bool> NoRestoreOption = CommonOptions.NoRestoreOption;
-        public readonly Option<bool> NoBuildOption = MicrosoftTestingPlatformOptions.NoBuildOption;
-        public readonly Option<bool> NoAnsiOption = MicrosoftTestingPlatformOptions.NoAnsiOption;
-        public readonly Option<bool> NoProgressOption = MicrosoftTestingPlatformOptions.NoProgressOption;
-        public readonly Option<OutputOptions> OutputOption = MicrosoftTestingPlatformOptions.OutputOption;
-        public readonly Option<string> ListTestsOption = MicrosoftTestingPlatformOptions.ListTestsOption;
-        public readonly Option<bool> NoLaunchProfileOption = MicrosoftTestingPlatformOptions.NoLaunchProfileOption;
-        public readonly Option<bool> NoLaunchProfileArgumentsOption = MicrosoftTestingPlatformOptions.NoLaunchProfileArgumentsOption;
-        public readonly Option<string[]> MTPTargetOption = MicrosoftTestingPlatformOptions.MTPTargetOption;
+
+        public readonly Option<bool> NoBuildOption = new("--no-build")
+        {
+            Description = CliCommandStrings.CmdNoBuildDescription
+        };
+
+        public readonly Option<bool> NoAnsiOption = new("--no-ansi")
+        {
+            Description = CliCommandStrings.CmdNoAnsiDescription,
+            Arity = ArgumentArity.Zero
+        };
+
+        public readonly Option<bool> NoProgressOption = new("--no-progress")
+        {
+            Description = CliCommandStrings.CmdNoProgressDescription,
+            Arity = ArgumentArity.Zero
+        };
+
+        public readonly Option<OutputOptions> OutputOption = new("--output")
+        {
+            Description = CliCommandStrings.CmdTestOutputDescription,
+            Arity = ArgumentArity.ExactlyOne
+        };
+
+        public const string ListTestsOptionName = "--list-tests";
+
+        public readonly Option<string> ListTestsOption = new(ListTestsOptionName)
+        {
+            Description = CliCommandStrings.CmdListTestsDescription,
+            Arity = ArgumentArity.Zero
+        };
+
+        public readonly Option<bool> NoLaunchProfileOption = new("--no-launch-profile")
+        {
+            Description = CliCommandStrings.CommandOptionNoLaunchProfileDescription,
+            Arity = ArgumentArity.Zero
+        };
+
+        public readonly Option<bool> NoLaunchProfileArgumentsOption = new("--no-launch-profile-arguments")
+        {
+            Description = CliCommandStrings.CommandOptionNoLaunchProfileArgumentsDescription
+        };
+
+        public readonly Option<string[]> MTPTargetOption = CommonOptions.CreateRequiredMSBuildTargetOption(CliConstants.MTPTarget);
 
         public ICustomHelp? CustomHelpLayoutProvider { get; set; }
 
@@ -157,34 +255,154 @@ internal abstract class TestCommandDefinition : Command
 
     public sealed class VSTest : TestCommandDefinition
     {
-        public readonly Option<string> SettingsOption = VSTestOptions.SettingsOption;
-        public readonly Option<bool> ListTestsOption = VSTestOptions.ListTestsOption;
+        public readonly Option<string> SettingsOption = new Option<string>("--settings", "-s")
+        {
+            Description = CliCommandStrings.CmdSettingsDescription,
+            HelpName = CliCommandStrings.CmdSettingsFile
+        }.ForwardAsSingle(o => $"-property:VSTestSetting={VSTestOptions.SurroundWithDoubleQuotes(CommandDirectoryContext.GetFullPath(o))}");
+
+        public readonly Option<bool> ListTestsOption = new Option<bool>("--list-tests", "-t")
+        {
+            Description = CliCommandStrings.CmdListTestsDescription,
+            Arity = ArgumentArity.Zero
+        }.ForwardAs("-property:VSTestListTests=true");
+
         public readonly Option<IReadOnlyDictionary<string, string>> TestEnvOption = CommonOptions.TestEnvOption;
-        public readonly Option<string> FilterOption = VSTestOptions.FilterOption;
-        public readonly Option<IEnumerable<string>> AdapterOption = VSTestOptions.AdapterOption;
-        public readonly Option<IEnumerable<string>> LoggerOption = VSTestOptions.LoggerOption;
-        public readonly Option<string> OutputOption = VSTestOptions.OutputOption;
+
+        public readonly Option<string> FilterOption = new Option<string>("--filter")
+        {
+            Description = CliCommandStrings.CmdTestCaseFilterDescription,
+            HelpName = CliCommandStrings.CmdTestCaseFilterExpression
+        }.ForwardAsSingle(o => $"-property:VSTestTestCaseFilter={VSTestOptions.SurroundWithDoubleQuotes(o!)}");
+
+        public readonly Option<IEnumerable<string>> AdapterOption = new Option<IEnumerable<string>>("--test-adapter-path")
+        {
+            Description = CliCommandStrings.CmdTestAdapterPathDescription,
+            HelpName = CliCommandStrings.CmdTestAdapterPath
+        }.ForwardAsSingle(o => $"-property:VSTestTestAdapterPath={VSTestOptions.SurroundWithDoubleQuotes(string.Join(";", o!.Select(CommandDirectoryContext.GetFullPath)))}")
+        .AllowSingleArgPerToken();
+
+        public readonly Option<IEnumerable<string>> LoggerOption = new Option<IEnumerable<string>>("--logger", "-l")
+        {
+            Description = CliCommandStrings.CmdLoggerDescription,
+            HelpName = CliCommandStrings.CmdLoggerOption
+        }.ForwardAsSingle(o =>
+        {
+            var loggersString = string.Join(";", VSTestOptions.GetSemiColonEscapedArgs(o!));
+            return $"-property:VSTestLogger={VSTestOptions.SurroundWithDoubleQuotes(loggersString)}";
+        })
+        .AllowSingleArgPerToken();
+
+        public readonly Option<string> OutputOption = new Option<string>("--output", "-o")
+        {
+            Description = CliCommandStrings.CmdOutputDescription,
+            HelpName = CliCommandStrings.TestCmdOutputDir
+        }
+        .ForwardAsOutputPath("OutputPath", true);
+
         public readonly Option<string> ArtifactsPathOption = CommonOptions.CreateArtifactsPathOption();
-        public readonly Option<string> DiagOption = VSTestOptions.DiagOption;
-        public readonly Option<bool> NoBuildOption = VSTestOptions.NoBuildOption;
-        public readonly Option<string> ResultsOption = VSTestOptions.ResultsOption;
-        public readonly Option<IEnumerable<string>> CollectOption = VSTestOptions.CollectOption;
-        public readonly Option<bool> BlameOption = VSTestOptions.BlameOption;
-        public readonly Option<bool> BlameCrashOption = VSTestOptions.BlameCrashOption;
-        public readonly Option<string> BlameCrashDumpOption = VSTestOptions.BlameCrashDumpOption;
-        public readonly Option<bool> BlameCrashAlwaysOption = VSTestOptions.BlameCrashAlwaysOption;
-        public readonly Option<bool> BlameHangOption = VSTestOptions.BlameHangOption;
-        public readonly Option<string> BlameHangDumpOption = VSTestOptions.BlameHangDumpOption;
-        public readonly Option<string> BlameHangTimeoutOption = VSTestOptions.BlameHangTimeoutOption;
-        public readonly Option<bool> NoLogoOption = VSTestOptions.NoLogoOption;
+
+        public readonly Option<string> DiagOption = new Option<string>("--diag", "-d")
+        {
+            Description = CliCommandStrings.CmdPathTologFileDescription,
+            HelpName = CliCommandStrings.CmdPathToLogFile
+        }
+        .ForwardAsSingle(o => $"-property:VSTestDiag={VSTestOptions.SurroundWithDoubleQuotes(CommandDirectoryContext.GetFullPath(o))}");
+
+        public readonly Option<bool> NoBuildOption = new Option<bool>("--no-build")
+        {
+            Description = CliCommandStrings.CmdNoBuildDescription,
+            Arity = ArgumentArity.Zero
+        }.ForwardAs("-property:VSTestNoBuild=true");
+
+        public readonly Option<string> ResultsOption = new Option<string>("--results-directory")
+        {
+            Description = CliCommandStrings.CmdResultsDirectoryDescription,
+            HelpName = CliCommandStrings.CmdPathToResultsDirectory
+        }.ForwardAsSingle(o => $"-property:VSTestResultsDirectory={VSTestOptions.SurroundWithDoubleQuotes(CommandDirectoryContext.GetFullPath(o))}");
+
+        public readonly Option<IEnumerable<string>> CollectOption = new Option<IEnumerable<string>>("--collect")
+        {
+            Description = CliCommandStrings.cmdCollectDescription,
+            HelpName = CliCommandStrings.cmdCollectFriendlyName
+        }.ForwardAsSingle(o => $"-property:VSTestCollect=\"{string.Join(";", VSTestOptions.GetSemiColonEscapedArgs(o!))}\"")
+        .AllowSingleArgPerToken();
+
+        public readonly Option<bool> BlameOption = new Option<bool>("--blame")
+        {
+            Description = CliCommandStrings.CmdBlameDescription,
+            Arity = ArgumentArity.Zero
+        }.ForwardIfEnabled("-property:VSTestBlame=true");
+
+        public readonly Option<bool> BlameCrashOption = new Option<bool>("--blame-crash")
+        {
+            Description = CliCommandStrings.CmdBlameCrashDescription,
+            Arity = ArgumentArity.Zero
+        }.ForwardIfEnabled("-property:VSTestBlameCrash=true");
+
+        public readonly Option<string> BlameCrashDumpOption = CreateBlameCrashDumpOption();
+
+        public readonly Option<bool> BlameCrashAlwaysOption = new Option<bool>("--blame-crash-collect-always")
+        {
+            Description = CliCommandStrings.CmdBlameCrashCollectAlwaysDescription,
+            Arity = ArgumentArity.Zero
+        }.ForwardIfEnabled(["-property:VSTestBlameCrash=true", "-property:VSTestBlameCrashCollectAlways=true"]);
+
+        public readonly Option<bool> BlameHangOption = new Option<bool>("--blame-hang")
+        {
+            Description = CliCommandStrings.CmdBlameHangDescription,
+            Arity = ArgumentArity.Zero
+        }.ForwardAs("-property:VSTestBlameHang=true");
+
+        public readonly Option<string> BlameHangDumpOption = CreateBlameHangDumpOption();
+
+        public readonly Option<string> BlameHangTimeoutOption = new Option<string>("--blame-hang-timeout")
+        {
+            Description = CliCommandStrings.CmdBlameHangTimeoutDescription,
+            HelpName = CliCommandStrings.HangTimeoutArgumentName
+        }.ForwardAsMany(o => ["-property:VSTestBlameHang=true", $"-property:VSTestBlameHangTimeout={o}"]);
+
+        public readonly Option<bool> NoLogoOption = CommonOptions.CreateNoLogoOption(forwardAs: "--property:VSTestNoLogo=true", description: CliCommandStrings.TestCmdNoLogo);
+
         public readonly Option<string> RuntimeOption = CommonOptions.CreateRuntimeOption(CliCommandStrings.TestRuntimeOptionDescription);
+
         public readonly Option<bool> NoRestoreOption = CommonOptions.NoRestoreOption;
+
         public readonly Option<bool> InteractiveOption = CommonOptions.CreateInteractiveMsBuildForwardOption();
+
         public readonly Option<string> ArchitectureOption = CommonOptions.ArchitectureOption;
+
         public readonly Option<string> OperatingSystemOption = CommonOptions.OperatingSystemOption;
+
         public readonly Option<ReadOnlyDictionary<string, string>?> PropertiesOption = CommonOptions.PropertiesOption;
+
         public readonly Option<bool> DisableBuildServersOption = CommonOptions.CreateDisableBuildServersOption();
-        public readonly Option<string[]> VsTestTargetOption = VSTestOptions.VsTestTargetOption;
+
+        public readonly Option<string[]> VsTestTargetOption = CommonOptions.CreateRequiredMSBuildTargetOption("VSTest");
+
+        private static Option<string> CreateBlameCrashDumpOption()
+        {
+            Option<string> result = new Option<string>("--blame-crash-dump-type")
+            {
+                Description = CliCommandStrings.CmdBlameCrashDumpTypeDescription,
+                HelpName = CliCommandStrings.CrashDumpTypeArgumentName,
+            }
+            .ForwardAsMany(o => ["-property:VSTestBlameCrash=true", $"-property:VSTestBlameCrashDumpType={o}"]);
+            result.AcceptOnlyFromAmong(["full", "mini"]);
+            return result;
+        }
+
+        private static Option<string> CreateBlameHangDumpOption()
+        {
+            Option<string> result = new Option<string>("--blame-hang-dump-type")
+            {
+                Description = CliCommandStrings.CmdBlameHangDumpTypeDescription,
+                HelpName = CliCommandStrings.HangDumpTypeArgumentName
+            }
+            .ForwardAsMany(o => ["-property:VSTestBlameHang=true", $"-property:VSTestBlameHangDumpType={o}"]);
+            result.AcceptOnlyFromAmong(["full", "mini", "none"]);
+            return result;
+        }
 
         public VSTest()
             : base(CliCommandStrings.DotnetTestCommandVSTestDescription)
